Ticket 586: ConcurrentModificationException when editing route
The cause for the exception was that when a message was handled, the list of all route definitions was looped over. If a route was udated/added/deleted during the time a message was handled, it meant that the same list that was iterated over was updated at the same time. This cause an exception to be thrown. To avoid this problem, the fix creates a copy of the route definitions list that is looped over, instead of the original one.