Ticket 668: Compositing route filter possibility
Possibility to filter in the compositing rules added.

Some redesign in CompositingRule was done while enabling the support for filtering. To ensure that correct information on previous period is collected (to check which sources and angles (in case of scans) are to be expected in this period) a different scheme for this had to be implemented. Now, complete file entry information has to be collected for the previous period and the filtering then applied on the result, to see which entries in the previous period matched the filters. This is however only done once for each period. After the first query, the information is stored in the CompositeTimerData.